requirements.txt
streamlit
pandas
mysql-connector-python
requests
plotly
python-dotenv
